+++ source/xtreme3d/engine.pas
@@ -100,9 +100,7 @@ begin
     result := elapsedMsec * 0.001;
 end;
 
-{
-function PtrToReal(p: pchar): real; cdecl;
+function PointerToReal(p: pchar): real; cdecl;
 begin
-    result := Integer(p);
+    result := real(NativeInt(p));
 end;
-}